Add files via upload
This research discusses the implementation of a social media-based tourism information system to improve the promotion of tourism destinations in Muara Enim, South Sumatra. With a focus on TikTok and Instagram, this research addresses the lack of understanding of Muara Enim's tourism potential among people outside the region through qualitative methods and data analysis. The results show that the utilization of social media, such as TikTok and Instagram, effectively increases the visibility of Muara Enim's tourism destinations, attracts tourists, and expands people's understanding of the tourism potential in the area. In the context of contribution, this research highlights the role of social media as an effective tool for local tourism development. As a suggestion for future research, further studies regarding the interaction and direct response of tourists to social media content, as well as evaluating the economic impact of increased tourism in Muara Enim. This can help a deeper understanding of the potential and challenges in developing the tourism sector through social media in Muara Enim.